requirements.txt
alembic==1.4.2
astroid==2.3.3
Babel==2.8.0
Click==7.0
decorator==4.4.1
dominate==2.5.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
Mako==1.1.2
MarkupSafe==1.1.1
mccabe==0.6.1
networkx==2.0
psycopg2-binary==2.8.4
pylint==2.4.4
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
python-dateutil==2.8.1
python-dotenv==0.12.0
python-editor==1.0.4
pytz==2019.3
six==1.13.0
SQLAlchemy==1.3.12
typed-ast==1.4.0
visitor==0.1.3
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
z3-solver==4.8.7.0
